refactor: move files to folders that better represent their use (#2564)
