+++ src/processors/PerformanceTrace.ts
