+++ src/tools/ToolDefinition.ts
@@ -11,7 +11,7 @@ import type {
   HeapSnapshotDetailedClassDiff,
   DuplicateStringGroup,
   HeapEdgesQueryOptions,
-} from '../HeapSnapshotManager.js';
+} from '../processors/HeapSnapshotManager.js';
 import type {McpPage} from '../McpPage.js';
 import {zod} from '../third_party/index.js';
 import type {
@@ -29,14 +29,17 @@ import type {
   Protocol,
   Page,
 } from '../third_party/index.js';
-import type {InsightName, TraceResult} from '../trace-processing/parse.js';
+import type {InsightName, TraceResult} from '../processors/PerformanceTrace.js';
 import type {
   TextSnapshotNode,
   GeolocationOptions,
   ExtensionServiceWorker,
 } from '../types.js';
 import type {PaginationOptions} from '../types.js';
-import type {WaitForEventsResult, DialogAction} from '../WaitForHelper.js';
+import type {
+  WaitForEventsResult,
+  DialogAction,
+} from '../utils/WaitForHelper.js';
 
 import type {ToolCategory} from './categories.js';
 import type {ToolGroups} from './thirdPartyDeveloper.js';
