+++ tests/formatters/ConsoleFormatter.test.ts
@@ -8,7 +8,7 @@ import {describe, it} from 'node:test';
 
 import {SymbolizedError} from '../../src/devtools/DevtoolsUtils.js';
 import {ConsoleFormatter} from '../../src/formatters/ConsoleFormatter.js';
-import {UncaughtError} from '../../src/PageCollector.js';
+import {UncaughtError} from '../../src/collectors/PageCollector.js';
 import type {ConsoleMessage, Protocol} from '../../src/third_party/index.js';
 import type {DevTools} from '../../src/third_party/index.js';
