+++ tests/utils/WaitForHelper.test.ts
@@ -7,8 +7,8 @@
 import assert from 'node:assert';
 import {describe, it} from 'node:test';
 
-import {serverHooks} from './server.js';
-import {html, withMcpContext} from './utils.js';
+import {serverHooks} from '../server.js';
+import {html, withMcpContext} from '../utils.js';
 
 describe('WaitForHelper', () => {
   const server = serverHooks();
